requirements.txt
fastapi
uvicorn
jinja2
pydantic
requests
python-dotenv
sqlite-utils
